FRONTEND bus delete function and viewing all bus done

+++ src/main/java/com/company/rll/repository/admin/BusRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BusRepository extends JpaRepository<BusEntity, Integer> {}
+public interface BusRepository extends JpaRepository<BusEntity, Long> {}
